Apply remaining parts of the unix port patch found at Fedora and Debian repositories
Authors: - Hans de Goede <hdegoede@redhat.com>,
         - Peter de Wachter <pdewacht@gmail.com>,
         - Peter Pentchev <roam@ringlet.net>

+++ src/shooter.h
@@ -83,8 +83,8 @@ typedef struct {
 	int difficulty;
 
 	// player related
-	long unsigned int score;
-	long unsigned int show_score;
+	unsigned int score;
+	unsigned int show_score;
 	int lives;
 	int power_gauge;
 	int power_level;
